ensure values aren't empty strings

+++ ddl/drop_column.go
@@ -1,6 +1,7 @@
 package ddl
 
 import (
+	"errors"
 	"fmt"
 
 	"github.com/orsinium-labs/sequel/dbconf"
@@ -16,6 +17,12 @@ func DropColumn(table Safe, col Safe) tDropColumn {
 }
 
 func (q tDropColumn) SQL(conf dbconf.Config) (string, error) {
+	if q.table == "" {
+		return "", errors.New("table name must not be empty")
+	}
+	if q.col == "" {
+		return "", errors.New("column name must not be empty")
+	}
 	sql := fmt.Sprintf("ALTER TABLE %s DROP COLUMN %s", q.table, q.col)
 	return sql, nil
 }
