+++ ddl/truncate_table.go
@@ -1,6 +1,7 @@
 package ddl
 
 import (
+	"errors"
 	"fmt"
 
 	"github.com/orsinium-labs/sequel/dbconf"
@@ -20,6 +21,10 @@ func TruncateTable(table Safe) tTruncateTable {
 func (q tTruncateTable) SQL(conf dbconf.Config) (string, error) {
 	// https://en.wikipedia.org/wiki/Data_definition_language#TRUNCATE_statement
 	// https://www.sqlite.org/lang_delete.html
+
+	if q.table == "" {
+		return "", errors.New("table name must not be empty")
+	}
 	prefix := "TRUNCATE TABLE"
 	if conf.Dialect == dbconf.SQLite {
 		prefix = "DELETE FROM"
